Employee Display take 2
I think it might be easier to restart rather than comb through the existing code, so,  these are the starting files, all blank at the moment except for employee-display.php, employee-display-scripts.php, and employee-display-cpt.php.  The plugin was successfully activated on a local WP install and when 'view-source-code' was accessed from the frontend, both the CSS and JavaScript files are being included.